requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# RPA自动化
selenium==4.15.2
webdriver-manager==4.0.1

# HTTP请求
requests==2.31.0
httpx==0.25.2

# 任务队列
celery==5.3.4
redis==5.0.1

# 日志
loguru==0.7.2

# 数据处理
pandas==2.1.3
numpy==1.25.2

# 文档解析
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2

# 配置管理
python-dotenv==1.0.0

# 时间处理
python-dateutil==2.8.2

# 类型检查
mypy==1.7.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 代码格式化
black==23.11.0
isort==5.12.0

# 安全
cryptography==41.0.7